+++ src/core/map_view.cpp
@@ -24,7 +24,7 @@
 #include <QRectF>
 
 #include "../map.h"
-#include "../template.h"
+#include "../templates/template.h"
 #include "../util.h"
 #include "../util/xml_stream_util.h"
