+++ src/fileformats/file_import_export.cpp
@@ -23,12 +23,12 @@
 
 #include <QFileInfo>
 
-#include "map.h"
-#include "symbol.h"
-#include "template.h"
-#include "object.h"
-#include "symbol_line.h"
-#include "symbol_point.h"
+#include "../map.h"
+#include "../symbol.h"
+#include "../templates/template.h"
+#include "../object.h"
+#include "../symbol_line.h"
+#include "../symbol_point.h"
 
 
 // ### ImportExport ###
