+++ src/fileformats/ocad8_file_format.cpp
@@ -35,22 +35,22 @@
 #include "core/map_view.h"
 #include "xml_file_format.h"
 #include "file_import_export.h"
-#include "map.h"
-#include "map_part.h"
-#include "object.h"
-#include "object_text.h"
-#include "settings.h"
-#include "symbol.h"
-#include "symbol_area.h"
-#include "symbol_combined.h"
-#include "symbol_line.h"
-#include "symbol_point.h"
-#include "symbol_text.h"
-#include "template.h"
-#include "template_image.h"
-#include "template_map.h"
-#include "util.h"
-#include "util/encoding.h"
+#include "../map.h"
+#include "../map_part.h"
+#include "../object.h"
+#include "../object_text.h"
+#include "../settings.h"
+#include "../symbol.h"
+#include "../symbol_area.h"
+#include "../symbol_combined.h"
+#include "../symbol_line.h"
+#include "../symbol_point.h"
+#include "../symbol_text.h"
+#include "../templates/template.h"
+#include "../templates/template_image.h"
+#include "../templates/template_map.h"
+#include "../util.h"
+#include "../util/encoding.h"
 
 // ### OCAD8FileFormat ###
