+++ src/fileformats/ocd_file_import.cpp
@@ -40,19 +40,19 @@
 #include "core/georeferencing.h"
 #include "ocad8_file_format.h"
 #include "ocad8_file_format_p.h"
-#include "map.h"
-#include "object_text.h"
-#include "settings.h"
-#include "symbol_area.h"
-#include "symbol_combined.h"
-#include "symbol_line.h"
-#include "symbol_point.h"
-#include "symbol_text.h"
-#include "template.h"
-#include "template_image.h"
-#include "template_map.h"
-#include "util.h"
-#include "util/encoding.h"
+#include "../map.h"
+#include "../object_text.h"
+#include "../settings.h"
+#include "../symbol_area.h"
+#include "../symbol_combined.h"
+#include "../symbol_line.h"
+#include "../symbol_point.h"
+#include "../symbol_text.h"
+#include "../templates/template.h"
+#include "../templates/template_image.h"
+#include "../templates/template_map.h"
+#include "../util.h"
+#include "../util/encoding.h"
 
 
 namespace {
