+++ src/fileformats/xml_file_format.cpp
@@ -39,18 +39,18 @@
 #include "core/map_printer.h"
 #include "core/map_view.h"
 #include "file_import_export.h"
-#include "map.h"
-#include "object.h"
-#include "object_text.h"
-#include "settings.h"
-#include "symbol_area.h"
-#include "symbol_combined.h"
-#include "symbol_line.h"
-#include "symbol_point.h"
-#include "symbol_text.h"
-#include "template.h"
-#include "undo_manager.h"
-#include "util/xml_stream_util.h"
+#include "../map.h"
+#include "../object.h"
+#include "../object_text.h"
+#include "../settings.h"
+#include "../symbol_area.h"
+#include "../symbol_combined.h"
+#include "../symbol_line.h"
+#include "../symbol_point.h"
+#include "../symbol_text.h"
+#include "../templates/template.h"
+#include "../undo_manager.h"
+#include "../util/xml_stream_util.h"
 
 // ### XMLFileFormat definition ###
