+++ src/gui/widgets/template_list_widget.cpp
@@ -43,11 +43,11 @@
 #include "../../map_widget.h"
 #include "../../object.h"
 #include "../../settings.h"
-#include "../../template.h"
-#include "../../template_adjust.h"
-#include "../../template_map.h"
-#include "../../template_position_dock_widget.h"
-#include "../../template_tool_move.h"
+#include "../../templates/template.h"
+#include "../../templates/template_adjust.h"
+#include "../../templates/template_map.h"
+#include "../../templates/template_position_dock_widget.h"
+#include "../../templates/template_tool_move.h"
 #include "../../util.h"
 #include "../../util/item_delegates.h"
 #include "../../util/scoped_signals_blocker.h"
