+++ src/map_editor.cpp
@@ -90,11 +90,11 @@
 #include "symbol_area.h"
 #include "symbol_dialog_replace.h"
 #include "symbol_point.h"
-#include "template.h"
-#include "template_dialog_reopen.h"
-#include "template_track.h"
-#include "template_position_dock_widget.h"
-#include "template_tool_paint.h"
+#include "templates/template.h"
+#include "templates/template_dialog_reopen.h"
+#include "templates/template_track.h"
+#include "templates/template_position_dock_widget.h"
+#include "templates/template_tool_paint.h"
 #include "tool.h"
 #include "tool_boolean.h"
 #include "tool_cut.h"
