+++ src/symbol_setting_dialog.cpp
@@ -44,8 +44,8 @@
 #include "gui/main_window.h"
 #include "map_editor.h"
 #include "map_widget.h"
-#include "template.h"
-#include "template_image.h"
+#include "templates/template.h"
+#include "templates/template_image.h"
 #include "symbol_point_editor.h"
 #include "symbol_combined.h"
 #include "symbol_properties_widget.h"
