+++ src/templates/template_adjust.cpp
