+++ src/templates/template_adjust.h
