+++ src/templates/template_dialog_reopen.cpp
