+++ src/templates/template_dialog_reopen.h
