+++ src/templates/template_image.cpp
