+++ src/templates/template_image.h
