+++ src/templates/template_map.cpp
