+++ src/templates/template_map.h
