+++ src/templates/template_tool_move.h
