+++ src/templates/template_tool_paint.cpp
