+++ src/templates/template_tool_paint.h
