+++ src/transformation.cpp
@@ -29,7 +29,7 @@
 #include "core/map_coord.h"
 #include "fileformats/file_format.h"
 #include "matrix.h"
-#include "template.h"
+#include "templates/template.h"
 #include "util/xml_stream_util.h"
 
 // ### PassPoint ###
