+++ test/file_format_t.cpp
@@ -31,7 +31,7 @@
 #include "../src/mapper_resource.h"
 #include "../src/object.h"
 #include "../src/settings.h"
-#include "../src/template.h"
+#include "../src/templates/template.h"
 #include "../src/undo_manager.h"
